+++ tests/test_cliutil.py
@@ -17,11 +17,26 @@ def test_SessionContext(testsdir):
         pass
 
 
-def test_bibtex2source():
-    bibtex2source(Record('book', 'id', author='M, R and G, H and Z, U'))
-    bibtex2source(Record('book', 'id', editor='M, R and G, H'))
-    bibtex2source(Record('book', 'id', title='tb', customfield='cf', year="1920}"))
-    assert bibtex2source(Record('misc', 'Id', title='title')).id == 'Id'
+@pytest.mark.parametrize(
+    'genre,id_,md,kw,assertion',
+    [
+        ('book', 'id', dict(author='M, R and G, H and Z, U'), None, None),
+        ('book', 'id', dict(author='M, R and G, H'), None, None),
+        ('book', 'id', dict(title='tb', customfield='cf', year="1920}"), None, None),
+        ('misc', 'Id', dict(title='title'), None, lambda src: src.id == 'Id'),
+        ('misc', 'Id', dict(title='title'), dict(lowercase_id=True), lambda src: src.id == 'id'),
+        (
+            'misc',
+            'a-b',
+            dict(url='https://example.org/~yurok/index.php'),
+            dict(sluggify_id=False, latex_unescape=False),
+            lambda src: src.id == 'a-b' and src.url == 'https://example.org/~yurok/index.php'),
+    ]
+)
+def test_bibtex2source(genre, id_, md, kw, assertion):
+    src = bibtex2source(Record(genre, id_, **md), **(kw or {}))
+    if assertion:
+        assert assertion(src)
 
 
 def test_Data(mocker):